# Go2Go.Model/ViewModels/DriverBalanceViewModel.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Go2Go.Model.ViewModels
{
    public class DriverBalanceViewModel
    {
        public int DriverId { get; set; }
        public decimal Balance { get; set; }
    }
}